+++ src/js/languages/ca.js
@@ -15,6 +15,9 @@
 		exports.ca = {
 			"mejs.plural-form": 1,
 
+			// core/mediaelement.js
+			"mejs.download-file": "Descarregar arxiu",
+
 			// renderers/flash.js
 			// "mejs.install-flash": "You are using a browser that does not have Flash player enabled or installed. Please turn on your Flash player plugin or download the latest version from https://get.adobe.com/flashplayer/",