+++ src/js/languages/de.js
@@ -16,6 +16,9 @@
 		exports.de = {
 			"mejs.plural-form": 1,
 
+			// core/mediaelement.js
+			"mejs.download-file": "Datei herunterladen",
+
 			// renderers/flash.js
 			"mejs.install-flash": "Ihr Browser unterstützt kein Flash. Bitte aktivieren Sie Flash bzw. laden Sie die aktuellste Flash-Version herunter unter https://get.adobe.com/flashplayer/",