+++ src/js/languages/en.js
@@ -14,6 +14,9 @@
 export const EN = {
 	"mejs.plural-form": 1,
 
+	// core/mediaelement.js
+	"mejs.download-file": "Download File",
+
 	// renderers/flash.js
 	"mejs.install-flash": "You are using a browser that does not have Flash player enabled or installed. Please turn on your Flash player plugin or download the latest version from https://get.adobe.com/flashplayer/",