+++ src/js/languages/pl.js
@@ -16,6 +16,9 @@
 		exports.pl = {
 			"mejs.plural-form": 9,
 
+			// core/mediaelement.js
+			"mejs.download-file": "Pobierz plik",
+
 			// renderers/flash.js
 			"mejs.install-flash": "Twoja przeglądarka nie ma włączonej lub zainstalowanej wtyczki Flash Player. Prosimy ją włączyć lub pobrać najnowszą wersję ze strony https://get.adobe.com/flashplayer/",