+++ src/js/utils/dom.js
@@ -50,7 +50,7 @@ export function fadeOut (el, duration = 400, callback) {
 	window.requestAnimationFrame(function animate(timestamp) {
 		start = start || timestamp;
 		const progress = timestamp - start;
-		el.style.opacity = 1 - progress / duration;
+		el.style.opacity = parseFloat(1 - progress / duration, 2);
 		if (progress > duration) {
 			if (callback && typeof(callback) === 'function') {
 				callback();
@@ -72,7 +72,7 @@ export function fadeIn (el, duration = 400, callback) {
 	window.requestAnimationFrame(function animate(timestamp) {
 		start = start || timestamp;
 		const progress = timestamp - start;
-		el.style.opacity = progress / duration;
+		el.style.opacity = parseFloat(progress / duration, 2);
 		if (progress > duration) {
 			if (callback && typeof(callback) === 'function') {
 				callback();